Prototype w/ Notes corresponding to LEDs
The code now functions with each note played by the CP being matched to a note that corresponds to an LED number and color. This will make for super easy integration of the "Recall Mode" where if the switch is on, the CP will not output a sound, just an LED flash (just to reiterate this is future work and has not been accomplished yet).
This code added some new mapping matrices where notes where mapped to LEDs and the frequency vector was transformed into a matrix to map each frequency to a note.